+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE_StateBackup/TOneWhS_BE_Bkup/Tables/SaleRate.sql
@@ -8,10 +8,12 @@
     [BED]              DATETIME        NOT NULL,
     [EED]              DATETIME        NULL,
     [SourceID]         VARCHAR (50)    NULL,
-    [Change]           TINYINT         NULL,
+    [Change]           INT             NULL,
     [StateBackupID]    BIGINT          NOT NULL,
     [LastModifiedTime] DATETIME        NULL
 );
+
+
